Move conversion scripts to scripts folder
