+++ scripts/compare_postprocessor_output.py
@@ -3,22 +3,22 @@ import difflib, argparse, subprocess, sys, os, multiprocessing, itertools
 
 def preprocess(compiler, file_contents):
     """ Get output from the preprocessing pass on a file.  """
-    output = subprocess.Popen(
+    return subprocess.Popen(
             [compiler, '-E', '-'],
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE,
             stdin=subprocess.PIPE).communicate(input=file_contents)[0]
-
-    def should_keep(line):
-        return str.strip(line) and line[0] != '#'
-
-    return filter(should_keep, output.splitlines())
 
 
 def preprocess_file(compiler, filename):
     """ Open a file and get the preprocessor output.  """
     with open(filename, 'rb') as f:
         return preprocess(compiler, f.read())
+
+
+def remove_empty_lines(text):
+    """ Remove empty lines from text.  """
+    return '\n'.join(filter(None, text.splitlines()))
 
 
 def file_contents_from_branch(filename, branch):
@@ -30,16 +30,21 @@ def file_contents_from_branch(filename, branch):
 def equal_to_file_on_branch(filename, branch, compiler):
     """
     Open a file on this branch and preprocess it.  Preprocess the same file
-    from another branch, and return a diff.
+    from another branch, and return whether the two files have (for all intents
+    and purposes) the same contents.
     """
     with open(filename, 'rb') as f:
         def p(text):
             return preprocess(compiler, text)
-        return difflib.unified_diff(p(f.read()),
-                p(file_contents_from_branch(filename, branch)),
-                fromfile=filename,
-                tofile=filename,
-                lineterm='')
+        return (p(f.read()) ==
+                p(file_contents_from_branch(filename, branch)))
+
+
+def process_single_file(filename, branch, compiler):
+    """ Like equal_to_file_on_branch, but also checks the file extension.  """
+    _, ext = os.path.splitext(filename)
+    return ((ext == '.h' or ext == '.cpp') and
+            not equal_to_file_on_branch(filename, branch, compiler))
 
 
 def is_source(filename):
@@ -53,9 +58,8 @@ def process(tup):
     Check a single file, and return its name if the check fails, otherwise
     return None.
     """
-    filename, branch, compiler = tup
-    failed = '\n'.join(equal_to_file_on_branch(filename, branch, compiler))
-    return failed if failed else None
+    failed = process_single_file(*tup)
+    return file if failed else None
 
 
 def main():
@@ -68,7 +72,7 @@ def main():
             '--branch', type=str, default='upstream/master',
             help='The branch to compare')
     parser.add_argument(
-            '--compiler', type=str, default='g++',
+            '--compiler', type=str, default='gcc',
             help='The compiler to use')
     args = parser.parse_args()
 
@@ -81,15 +85,10 @@ def main():
             itertools.cycle([args.branch]),
             itertools.cycle([args.compiler]))
 
-    pool = multiprocessing.Pool(10)
-
-    results = filter(None, pool.map(process, zipped))
-
-    pool.close()
-    pool.join()
+    results = filter(None, multiprocessing.Pool(10).map(process, zipped))
 
     if results:
-        print('\n\n'.join(results))
+        print('\n'.join(results))
         return 1
 
     return 0
